feat: introduce resource loader interface
